requirements.txt
pandas
matplotlib
seaborn
jupyter
scikit-learn
streamlit
plotly
python-dotenv
aiohttp
tqdm
pycountry